Popup Action Hotfix
dialog_active is set to true before the popup occurs, if there is dialogue in the Interactable that then triggers the popup action. This prevented the popup from occurring since it would check if sequence_active OR dialog_active was true. But dialog_active doesn't make much sense so yeah. Easy fix.